feat: migrate to Flarum 2.0
- Update dependencies (flarum/core ^2.0.0-beta, flarum-webpack-config ^3.0.0, flarum-tsconfig ^2.0.0)
- Migrate admin settings to Extend.Admin() extender pattern
- Convert to string-based extend() for PostStream
- Replace @control-color LESS variable with CSS custom property
- Update CI workflows for 2.x
