+++ .github/workflows/frontend.yml
@@ -4,12 +4,13 @@ on: [workflow_dispatch, push, pull_request]
 
 jobs:
   run:
-    uses: flarum/framework/.github/workflows/REUSABLE_frontend.yml@main
+    uses: flarum/framework/.github/workflows/REUSABLE_frontend.yml@2.x
     with:
       enable_bundlewatch: false
       enable_prettier: true
       enable_typescript: true
 
+      node_version: 18
       frontend_directory: ./js
       backend_directory: .
       js_package_manager: npm
