+++ composer.json
@@ -8,7 +8,8 @@
     "type": "flarum-extension",
     "license": "MIT",
     "require": {
-        "flarum/core": "^1.2.0",
+        "php": "^8.2",
+        "flarum/core": "^2.0.0-beta",
         "fof/gamification": "*"
     },
     "authors": [
