+++ js/package.json
@@ -3,12 +3,12 @@
     "private": true,
     "version": "0.0.0",
     "devDependencies": {
-        "flarum-webpack-config": "^2.0.0",
+        "flarum-webpack-config": "^3.0.0",
         "webpack": "^5.94.0",
         "webpack-cli": "^4.9.1",
         "prettier": "^2.5.1",
         "@flarum/prettier-config": "^1.0.0",
-        "flarum-tsconfig": "^1.0.2",
+        "flarum-tsconfig": "^2.0.0",
         "typescript": "^4.5.4"
     },
     "scripts": {
