+++ js/src/forum/index.ts
@@ -1,11 +1,10 @@
 import app from 'flarum/forum/app';
 import { extend } from 'flarum/common/extend';
-import PostStream from 'flarum/forum/components/PostStream';
 import MoreDiscussions from './components/MoreDiscussions';
 import insertAfter from './helpers/insertAfter';
 
 app.initializers.add('datlechin/flarum-more-discussions', () => {
-  extend(PostStream.prototype, 'oncreate', function () {
+  extend('flarum/forum/components/PostStream', 'oncreate', function () {
     const postStream: HTMLElement | null = document.querySelector('.PostStream');
     if (!postStream) return;
