+++ less/forum.less
@@ -2,7 +2,7 @@
   border-top: 1px solid #e5e5e5;
   margin-top: 20px;
   &-title {
-    color: @control-color;
+    color: var(--control-color);
     @media @phone {
       font-size: 18px;
       padding: 0 20px;
